+++ src/sdk/PnP.Core/Model/SharePoint/Core/Internal/ActivityIncomplete.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace PnP.Core.Model.SharePoint
+{
+    internal class ActivityIncomplete : IActivityIncomplete
+    {
+        public DateTime MissingDataBeforeDateTime { get; set; }
+
+        public bool WasThrottled { get; set; }
+
+        public bool ResultsPending { get; set; }
+
+        public bool NotSupported { get; set; }
+    }
+}